restructure fs layout
