+++ contracts/optimistic-ethereum/OVM/bridge/assets/OVM_L1ERC20Gateway.sol
@@ -5,13 +5,13 @@ pragma solidity >0.5.0 <0.8.0;
 pragma experimental ABIEncoderV2;
 
 /* Interface Imports */
-import { iOVM_L1ERC20Gateway } from "../../../iOVM/bridge/ERC20/iOVM_L1ERC20Gateway.sol";
-import { iOVM_L2ERC20Gateway } from "../../../iOVM/bridge/ERC20/iOVM_L2ERC20Gateway.sol";
-import { iAbs_BaseCrossDomainMessenger } from "../../../iOVM/bridge/iAbs_BaseCrossDomainMessenger.sol";
+import { iOVM_L1ERC20Gateway } from "../../../iOVM/bridge/assets/iOVM_L1ERC20Gateway.sol";
+import { iOVM_L2ERC20Gateway } from "../../../iOVM/bridge/assets/iOVM_L2ERC20Gateway.sol";
+import { iAbs_BaseCrossDomainMessenger } from "../../../iOVM/bridge/base/iAbs_BaseCrossDomainMessenger.sol";
 import { iOVM_ERC20 } from "../../../iOVM/precompiles/iOVM_ERC20.sol";
 
 /* Library Imports */
-import { OVM_CrossChainEnabled } from "../OVM_CrossChainEnabled.sol";
+import { OVM_CrossChainEnabled } from "../../../libraries/bridge/OVM_CrossChainEnabled.sol";
 
 /**
  * @title OVM_L1ERC20Gateway
