+++ contracts/optimistic-ethereum/OVM/bridge/assets/OVM_L2ERC20Gateway.sol
@@ -3,13 +3,15 @@ pragma solidity >0.5.0 <0.8.0;
 pragma experimental ABIEncoderV2;
 
 /* Interface Imports */
-import { iOVM_L2ERC20Gateway } from "../../../iOVM/bridge/ERC20/iOVM_L2ERC20Gateway.sol";
-import { iOVM_L1ERC20Gateway } from "../../../iOVM/bridge/ERC20/iOVM_L1ERC20Gateway.sol";
-import { iOVM_L2CrossDomainMessenger } from "../../../iOVM/bridge/iOVM_L2CrossDomainMessenger.sol";
+import { iOVM_L2ERC20Gateway } from "../../../iOVM/bridge/assets/iOVM_L2ERC20Gateway.sol";
+import { iOVM_L1ERC20Gateway } from "../../../iOVM/bridge/assets/iOVM_L1ERC20Gateway.sol";
+import { iOVM_L2CrossDomainMessenger } from "../../../iOVM/bridge/base/iOVM_L2CrossDomainMessenger.sol";
 
 /* Contract Imports */
-import { OVM_CrossChainEnabled } from "../OVM_CrossChainEnabled.sol";
-import { UniswapV2ERC20 } from "../../precompiles/UniswapV2ERC20.sol";
+import { UniswapV2ERC20 } from "../../../libraries/standards/UniswapV2ERC20.sol";
+
+/* Library Imports */
+import { OVM_CrossChainEnabled } from "../../../libraries/bridge/OVM_CrossChainEnabled.sol";
 
 /**
  * @title OVM_L2ERC20Gateway
