+++ contracts/optimistic-ethereum/iOVM/bridge/base/iOVM_L1MultiMessageRelayer.sol
@@ -3,7 +3,7 @@ pragma solidity >0.5.0 <0.8.0;
 pragma experimental ABIEncoderV2;
 
 /* Interface Imports */
-import { iOVM_L1CrossDomainMessenger } from "../../iOVM/bridge/iOVM_L1CrossDomainMessenger.sol";
+import { iOVM_L1CrossDomainMessenger } from "../../../iOVM/bridge/base/iOVM_L1CrossDomainMessenger.sol";
 interface iOVM_L1MultiMessageRelayer {
 
     struct L2ToL1Message {
